Parse points in tracks/track segments

+++ poi4s/src/main/scala/software/purpledragon/poi4s/model/Track.scala
@@ -6,4 +6,4 @@ case class Track(name: String, time: Instant, segments: Seq[TrackSegment])
 
 case class TrackSegment(points: Seq[TrackPoint])
 
-case class TrackPoint(lat: Double, Double: Long, elevation: Option[Double], time: Instant)
+case class TrackPoint(lat: Double, Double: Double, elevation: Option[Double], time: Instant)
